Place order api added

+++ src/main/java/com/example/Swiggato/model/OrderEntity.java
@@ -27,7 +27,7 @@ public class OrderEntity {
 
     String orderId; //UUID
 
-    int orderTotal;
+    double orderTotal;
 
     @CreationTimestamp
     Date orderTime;
